+++ backend/canisters/group/impl/src/lib.rs
@@ -776,14 +776,11 @@ impl Data {
 
         match initiator {
             BotInitiator::Command(command) => {
-                let granted_to_bot = if let Some(autonomous_permissions) = &bot.autonomous_permissions {
-                    &BotPermissions::union(&bot.permissions, autonomous_permissions)
-                } else {
-                    &bot.permissions
-                };
-
-                self.get_user_permissions(&command.initiator)
-                    .map(|u| BotPermissions::intersect(granted_to_bot, &u))
+                let user_permissions = self.get_user_permissions(&command.initiator)?;
+                Some(BotPermissions::union(
+                    &BotPermissions::intersect(&bot.permissions, &user_permissions),
+                    &bot.autonomous_permissions.clone().unwrap_or_default(),
+                ))
             }
             BotInitiator::Autonomous => bot.autonomous_permissions.clone(),
         }