+++ backend/canisters/user/impl/src/lib.rs
@@ -648,11 +648,7 @@ impl Data {
         // Get the granted permissions when initiated by command or API key
         let granted = match initiator {
             BotInitiator::Command(_) => {
-                if let Some(autonomous_permissions) = &bot.autonomous_permissions {
-                    &BotPermissions::union(&bot.permissions, autonomous_permissions)
-                } else {
-                    &bot.permissions
-                }
+                &BotPermissions::union(&bot.permissions, &bot.autonomous_permissions.clone().unwrap_or_default())
             }
             BotInitiator::Autonomous => match bot.autonomous_permissions.as_ref() {
                 Some(permissions) => permissions,